+++ CHANGES.md
@@ -5,6 +5,11 @@ Changes in 5.2
 --------------
 - Add $max_file_size param & max file size validation in Inputs.php & warehouse.js
 - Add Organization radio inputs in FirstLogin.fnc.php
+- Add db_sql_filter() function in database.inc.php
+- Add db_trans_rollback() function in database.inc.php
+- Deprecate $connection param for db_trans_start(), db_trans_commit() functions in database.inc.php
+- Remove $connection param for db_trans_query() in database.inc.php
+- Add $show_error optional param to db_query() & db_trans_query() in database.inc.php
 
 Changes in 5.1.1
 ----------------